fixed linting warnings, updated the dots menu in the CommentsMainContent.vue
