+++ Frontend/src/components/login/RegisterComponent.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import IconX from "@/components/icons/IconX.vue";
 
 const emit = defineEmits(['navigate'])
 </script>
